fix: conflict adjustment
