+++ src/components/modals/PaymentCapacityModal/index.tsx
@@ -21,7 +21,21 @@ import { currencyFormat } from "@utils/formatData/currency";
 
 import { incomeModalConfig } from "./IcomeModalConfig";
 import { StyledContainerClose, StyledModal } from "./styles";
-import { PaymentCapacityProps } from "./interface";
+
+
+
+export interface PaymentCapacityProps {
+  title: string;
+  handleClose: () => void;
+  portalId?: string;
+  reportedIncomeSources: number;
+  reportedFinancialObligations: number;
+  subsistenceReserve: number;
+  availableForNewCommitments: number;
+  maxVacationTerm: number;
+  maxAmount: number;
+}
+
 
 export const PaymentCapacity = (props: PaymentCapacityProps) => {
   const {
