+++ src/components/modals/PaymentCapacityModal/stories/props.ts
@@ -1,5 +1,5 @@
 import { ArgTypes } from "@storybook/react";
-import { PaymentCapacityProps } from '../interface';
+import { PaymentCapacityProps } from '..';
 
 export const props: Partial<ArgTypes<PaymentCapacityProps>> = {
   title: {
